+++ Runtime/Components/GlyphRichTextFormatter.cs
@@ -1,6 +1,5 @@
 using JetBrains.Annotations;
 using Rewired;
-using System;
 using System.Text;
 using System.Text.RegularExpressions;
 using TMPro;
@@ -28,22 +27,22 @@ namespace LMirman.RewiredGlyphs.Components
 		[Header("Update Contexts")]
 		[SerializeField]
 		[Tooltip("When enabled the text will be formatted in start.\n\n" +
-				 "This should usually be enabled unless you are setting formatted text elsewhere during initialization.")]
+		         "This should usually be enabled unless you are setting formatted text elsewhere during initialization.")]
 		private bool formatTextOnStart = true;
 		[SerializeField]
 		[Tooltip("When enabled the text will be updated with new sprites when glyphs are rebuilt.\n\n" +
-				 "This should usually be enabled so you are always showing up to date glyphs. However, if you do not want the value of text mesh to change consider disabling this.")]
+		         "This should usually be enabled so you are always showing up to date glyphs. However, if you do not want the value of text mesh to change consider disabling this.")]
 		private bool updateTextOnRebuildGlyphs = true;
 		[SerializeField]
 		[Tooltip("When enabled will check for text changes every frame. When text changes will automatically reformat text.\n\n" +
-				 "This is somewhat expensive and may lose glyph references if you append text so only enable if you don't want to or can't use the \"SetText\" method of this class.")]
+		         "This is somewhat expensive and may lose glyph references if you append text so only enable if you don't want to or can't use the \"SetText\" method of this class.")]
 		private bool automaticallyCheckForTextChanges;
 
 		[Header("Output Customization")]
 		[SerializeField]
 		[Tooltip("When enabled the glyph rich text will be replaced with an TMP inline sprite instead of text." +
-				 "If there is no sprite available for the glyph will fallback to description.\n\n" +
-				 "When disabled: Will always show the text description of the Glyph instead of its sprite.")]
+		         "If there is no sprite available for the glyph will fallback to description.\n\n" +
+		         "When disabled: Will always show the text description of the Glyph instead of its sprite.")]
 		private bool useSpritesWhenAvailable = true;
 		[SerializeField]
 		[Tooltip("The format for Glyph descriptions that replace glyph rich text values. Default: \"[{0}]\" will output \"[R2]\" for a primary fire action for example.")]
@@ -135,6 +134,31 @@ namespace LMirman.RewiredGlyphs.Components
 
 		/// <summary>
 		/// Replace glyph rich text such as &lt;glyph "Jump"&gt; in the provided <paramref name="text"/> with a sprite or description for its corresponding Glyph in the Rewired Glyph system.
+		/// <br/><br/>
+		/// <b>Syntax Rules:</b><br/>
+		/// - The glyph tag begins with verbatim `&lt;glyph ` and ends at any `&gt;` character.<br/>
+		///	- The first argument must <b>always</b> represent the action to represent as either an integer for the action ID or a string for the action name <i>if and only if</i> it does not contain any spaces.<br/>
+		/// - Following the action argument all arguments may appear in any order.<br/>
+		/// - Arguments are delimited by any occurrence of the space character, <i>even if the space is contained within quotes</i>.<br/>
+		/// <br/>
+		/// <b>Additional Arguments:</b><br/>
+		/// Following the action argument you may optionally include the following arguments to control the output glyph.
+		/// <br/><br/>
+		/// - `Player ID` (Default: 0)<br/>
+		/// Specify the player you'd like to show glyph for using `player=2` where 2 can be any int.
+		/// If an integer is alone without any specifier it is interpret as player id, such as `&lt;glyph Jump 2&gt;` where 2 is player id
+		/// <br/><br/>
+		/// - `Polarity` (Default: Positive)<br/>
+		/// Specify the expected axis or pole input for the action such as positive for "Move Right" and negative for "Move Left" on a Move Horizontal action
+		/// using `pole=Positive`, `pole=Negative`, or `pole=FullAxis`. You may exclude `pole` specifier from your input.
+		/// <br/><br/>
+		/// - `Controller Type` (Default: Current)<br/>
+		/// Specify the controller type to show symbols for using `type=Joystick`.
+		/// Valid values are `type=Current`, `type=Keyboard`, `type=Mouse`, and `type=Joystick`.<br/>
+		/// Warning: Specifying a controller that the runtime device doesn't have connected (such as Joystick) will show the `UNBOUND` symbol until they connect that controller type.<br/>
+		/// Warning: Specifying a controller that does not have the specified action bound to that specific controller will show the `UNBOUND` symbol without falling back to another controller.
+		/// <br/><br/>
+		/// All additional arguments that are string based are <b>case-insensitive</b> such that "TYPE=JOYSTICK" and "type=joystick" are equivalent inputs.
 		/// </summary>
 		/// <remarks>Does not mutate the provided string.</remarks>
 		[Pure]
@@ -150,51 +174,18 @@ namespace LMirman.RewiredGlyphs.Components
 				}
 
 				string[] splitArgs = match.Groups[1].Value.Split(' ');
-				bool didSetActionId = false;
-				int actionId = 0;
-				int player = 0;
-				Pole pole = Pole.Positive;
-				bool forceAxis = false;
-				foreach (string splitArg in splitArgs)
+				GlyphParseResult result = new GlyphParseResult(splitArgs);
+				Glyph glyph;
+				AxisRange axisRange;
+				if (result.useCurrentController)
 				{
-					if (string.IsNullOrWhiteSpace(splitArg))
-					{
-						continue;
-					}
-
-					string trimmedArg = splitArg.Trim('\"', '/', '\\');
-					// The first arg is always interpret as the action id.
-					if (!didSetActionId)
-					{
-						bool isInt = int.TryParse(trimmedArg, out int foundActionId);
-						// If the value is an int, we set it directly as the action id.
-						// Otherwise, try to find an action id with the string value from rewired
-						actionId = isInt ? foundActionId : ReInput.mapping.GetActionId(trimmedArg);
-						didSetActionId = true;
-						continue;
-					}
-
-					// We are no longer evaluating the first arg (since action id has been set)
-
-					if (int.TryParse(trimmedArg, out int playerId))
-					{
-						player = playerId;
-					}
-					else if (trimmedArg.Equals("Positive", StringComparison.OrdinalIgnoreCase))
-					{
-						pole = Pole.Positive;
-					}
-					else if (trimmedArg.Equals("Negative", StringComparison.OrdinalIgnoreCase))
-					{
-						pole = Pole.Negative;
-					}
-					else if (trimmedArg.Equals("Full", StringComparison.OrdinalIgnoreCase) || trimmedArg.Equals("Full Axis", StringComparison.OrdinalIgnoreCase))
-					{
-						forceAxis = true;
-					}
+					glyph = InputGlyphs.GetCurrentGlyph(result.actionId, result.pole, out axisRange, result.playerId, result.forceAxis);
+				}
+				else
+				{
+					glyph = InputGlyphs.GetGlyph(result.controllerType, result.actionId, result.pole, out axisRange, result.playerId, result.forceAxis);
 				}
 
-				Glyph glyph = InputGlyphs.GetCurrentGlyph(actionId, pole, out AxisRange axisRange, player, forceAxis);
 				Sprite sprite = glyph.GetSprite(axisRange);
 				if (useSpritesWhenAvailable && sprite != null && !glyph.PreferDescription)
 				{